correct multisig test file names
