+++ test/functional/system/4.X.multisignature_validated.js
@@ -15,434 +15,134 @@
 'use strict';
 
 var lisk = require('lisk-js');
-var accountFixtures = require('../../fixtures/accounts');
 var randomUtil = require('../../common/utils/random');
-var normalizer = require('../../common/utils/normalizer');
+var Scenarios = require('../common/scenarios');
+var transactionTypes = require('../../../helpers/transaction_types.js');
 var localCommon = require('./common');
 
-describe('system test (type 4) - effect of multisignature registration on memory tables', () => {
+describe('system test (type 4) - checking registered multisignature transaction against other transaction types', () => {
 	var library;
-	var multisigSender;
 
-	var multisigAccount = randomUtil.account();
-	var multisigTransaction;
-	var creditTransaction = lisk.transaction.createTransaction(
-		multisigAccount.address,
-		1000 * normalizer,
-		accountFixtures.genesis.password
+	var scenarios = {
+		regular: new Scenarios.Multisig(),
+	};
+
+	scenarios.regular.dapp = randomUtil.application();
+	var dappTransaction = lisk.dapp.createDapp(
+		scenarios.regular.account.password,
+		null,
+		scenarios.regular.dapp
 	);
-	var signer1 = randomUtil.account();
-	var signer2 = randomUtil.account();
+	scenarios.regular.dapp.id = dappTransaction.id;
 
-	localCommon.beforeBlock('system_4_multisig_account', lib => {
+	scenarios.regular.multiSigTransaction.ready = true;
+	scenarios.regular.multiSigTransaction.signatures = [];
+
+	scenarios.regular.members.map(member => {
+		var signature = lisk.multisignature.signTransaction(
+			scenarios.regular.multiSigTransaction,
+			member.password
+		);
+		scenarios.regular.multiSigTransaction.signatures.push(signature);
+	});
+
+	localCommon.beforeBlock('system_4_X_multisig_validated', lib => {
 		library = lib;
 	});
 
 	before(done => {
-		localCommon.addTransactionsAndForge(library, [creditTransaction], () => {
-			library.logic.account.get(
-				{ address: multisigAccount.address },
-				(err, sender) => {
-					multisigSender = sender;
+		localCommon.addTransactionsAndForge(
+			library,
+			[scenarios.regular.creditTransaction],
+			() => {
+				localCommon.addTransactionsAndForge(library, [dappTransaction], () => {
+					done();
+				});
+			}
+		);
+	});
+
+	it('adding to pool multisignature registration should be ok', done => {
+		localCommon.addTransaction(
+			library,
+			scenarios.regular.multiSigTransaction,
+			(err, res) => {
+				expect(res).to.equal(scenarios.regular.multiSigTransaction.id);
+				done();
+			}
+		);
+	});
+
+	describe('after forging one block', () => {
+		before(done => {
+			localCommon.forge(library, () => {
+				done();
+			});
+		});
+
+		it('transaction should be included', done => {
+			var filter = {
+				id: scenarios.regular.multiSigTransaction.id,
+			};
+			localCommon.getTransactionFromModule(library, filter, (err, res) => {
+				expect(err).to.be.null;
+				expect(res)
+					.to.have.property('transactions')
+					.which.is.an('Array');
+				expect(res.transactions.length).to.equal(1);
+				expect(res.transactions[0].id).to.equal(
+					scenarios.regular.multiSigTransaction.id
+				);
+				done();
+			});
+		});
+
+		it('adding to pool multisignature registration for same account should fail', done => {
+			localCommon.addTransaction(
+				library,
+				scenarios.regular.multiSigTransaction,
+				err => {
+					expect(err).to.equal('Account already has multisignatures enabled');
 					done();
 				}
 			);
 		});
-	});
 
-	describe('forge block with multisignature transaction', () => {
-		before('forge block with multisignature transaction', done => {
-			var keysgroup = [`+${signer1.publicKey}`, `+${signer2.publicKey}`];
-
-			multisigTransaction = lisk.multisignature.createMultisignature(
-				multisigAccount.password,
-				null,
-				keysgroup,
-				4,
-				2
-			);
-			var sign1 = lisk.multisignature.signTransaction(
-				multisigTransaction,
-				signer1.password
-			);
-			var sign2 = lisk.multisignature.signTransaction(
-				multisigTransaction,
-				signer2.password
-			);
-
-			multisigTransaction.signatures = [sign1, sign2];
-			multisigTransaction.ready = true;
-			localCommon.addTransactionsAndForge(library, [multisigTransaction], done);
-		});
-
-		describe('check sender db rows', () => {
-			var accountRow;
-
-			before(
-				'get mem_account, mem_account2multisignature and mem_account2u_multisignature rows',
-				() => {
-					return localCommon
-						.getAccountFromDb(library, multisigAccount.address)
-						.then(res => {
-							accountRow = res;
-						});
+		describe('adding to pool other transaction types from the same account', () => {
+			Object.keys(transactionTypes).forEach((key, index) => {
+				if (key === 'IN_TRANSFER' || key === 'OUT_TRANSFER') {
+					it(`type ${index}: ${key} should be rejected`, done => {
+						localCommon.loadTransactionType(
+							key,
+							scenarios.regular.account,
+							scenarios.regular.dapp,
+							true,
+							transaction => {
+								localCommon.addTransaction(library, transaction, err => {
+									expect(err).to.equal(
+										`Transaction type ${transaction.type} is frozen`
+									);
+									done();
+								});
+							}
+						);
+					});
+				} else if (key != 'MULTI') {
+					it(`type ${index}: ${key} should be ok`, done => {
+						localCommon.loadTransactionType(
+							key,
+							scenarios.regular.account,
+							scenarios.regular.dapp,
+							true,
+							transaction => {
+								localCommon.addTransaction(library, transaction, (err, res) => {
+									expect(res).to.equal(transaction.id);
+									done();
+								});
+							}
+						);
+					});
 				}
-			);
-
-			it('should include rows in mem_accounts2multisignatures', () => {
-				var signKeysInDb = _.map(
-					accountRow.mem_accounts2multisignatures,
-					row => {
-						return row.dependentId;
-					}
-				);
-				return expect(signKeysInDb).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should set multimin field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should set multilifetime field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-
-			it('should include rows in mem_accounts2u_multisignatures', () => {
-				var signKeysInDb = _.map(
-					accountRow.mem_accounts2u_multisignatures,
-					row => {
-						return row.dependentId;
-					}
-				);
-				return expect(signKeysInDb).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should set u_multimin field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.u_multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should set u_multilifetime field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.u_multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-		});
-
-		describe('check sender account', () => {
-			var account;
-
-			before('get multisignature account', done => {
-				library.logic.account.get(
-					{ address: multisigAccount.address },
-					(err, res) => {
-						expect(err).to.be.null;
-						account = res;
-						done();
-					}
-				);
-			});
-
-			it('should have multisignatures field set on account', () => {
-				return expect(account.multisignatures).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should have multimin field set on account', () => {
-				return expect(account.multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should have multilifetime field set on account', () => {
-				return expect(account.multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-
-			it('should have u_multisignatures field set on account', () => {
-				return expect(account.u_multisignatures).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should have u_multimin field set on account', () => {
-				return expect(account.u_multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should have u_multilifetime field set on account', () => {
-				return expect(account.u_multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-		});
-
-		describe('after deleting block', () => {
-			before('delete last block', done => {
-				library.modules.blocks.lastBlock.get();
-				library.modules.blocks.chain.deleteLastBlock(done);
-			});
-
-			describe('sender db rows', () => {
-				var accountRow;
-
-				before(
-					'get mem_account, mem_account2multisignature and mem_account2u_multisignature rows',
-					() => {
-						return localCommon
-							.getAccountFromDb(library, multisigAccount.address)
-							.then(res => {
-								accountRow = res;
-							});
-					}
-				);
-
-				it('should have no rows in mem_accounts2multisignatures', () => {
-					return expect(accountRow.mem_accounts2multisignatures).to.eql([]);
-				});
-
-				it('should have multimin field set to 0 on mem_accounts', () => {
-					return expect(accountRow.mem_accounts.multimin).to.eql(0);
-				});
-
-				it('should have multilifetime field set to 0 on mem_accounts', () => {
-					return expect(accountRow.mem_accounts.multilifetime).to.eql(0);
-				});
-
-				it('should have no rows in mem_accounts2u_multisignatures', () => {
-					return expect(accountRow.mem_accounts2u_multisignatures).to.eql([]);
-				});
-
-				it('should have u_multimin field set to 0 on mem_accounts', () => {
-					return expect(accountRow.mem_accounts.u_multimin).to.eql(0);
-				});
-
-				it('should have multilifetime field to 0 on mem_accounts', () => {
-					return expect(accountRow.mem_accounts.u_multilifetime).to.eql(0);
-				});
-			});
-
-			describe('sender account', () => {
-				var account;
-
-				before('get multisignature account', done => {
-					library.logic.account.get(
-						{ address: multisigAccount.address },
-						(err, res) => {
-							expect(err).to.be.null;
-							account = res;
-							done();
-						}
-					);
-				});
-
-				it('should set multisignatures field to null on account', () => {
-					return expect(account.multisignatures).to.be.null;
-				});
-
-				it('should set multimin field to 0 on account', () => {
-					return expect(account.multimin).to.eql(0);
-				});
-
-				it('should set multilifetime field to 0 on account', () => {
-					return expect(account.multilifetime).to.eql(0);
-				});
-
-				it('should set u_multisignatures field to null on account', () => {
-					return expect(account.u_multisignatures).to.be.null;
-				});
-
-				it('should set u_multimin field to null on account', () => {
-					return expect(account.u_multimin).to.eql(0);
-				});
-
-				it('should set u_multilifetime field to null on account', () => {
-					return expect(account.u_multilifetime).to.eql(0);
-				});
-			});
-		});
-	});
-
-	describe('apply unconfirmed transaction', () => {
-		before('apply unconfirmed multisig transaction', done => {
-			var keysgroup = [`+${signer1.publicKey}`, `+${signer2.publicKey}`];
-
-			multisigTransaction = lisk.multisignature.createMultisignature(
-				multisigAccount.password,
-				null,
-				keysgroup,
-				4,
-				2
-			);
-			var sign1 = lisk.multisignature.signTransaction(
-				multisigTransaction,
-				signer1.password
-			);
-			var sign2 = lisk.multisignature.signTransaction(
-				multisigTransaction,
-				signer2.password
-			);
-
-			multisigTransaction.signatures = [sign1, sign2];
-			multisigTransaction.ready = true;
-
-			library.logic.transaction.applyUnconfirmed(
-				multisigTransaction,
-				multisigSender,
-				done
-			);
-		});
-
-		describe('check sender db rows', () => {
-			var accountRow;
-
-			before(
-				'get mem_account, mem_account2multisignature and mem_account2u_multisignature rows',
-				() => {
-					return localCommon
-						.getAccountFromDb(library, multisigAccount.address)
-						.then(res => {
-							accountRow = res;
-						});
-				}
-			);
-
-			it('should have no rows in mem_accounts2multisignatures', () => {
-				return expect(accountRow.mem_accounts2multisignatures).to.eql([]);
-			});
-
-			it('should have multimin field set to 0 on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.multimin).to.eql(0);
-			});
-
-			it('should have multilifetime field set to 0 on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.multilifetime).to.eql(0);
-			});
-
-			it('should include rows in mem_accounts2u_multisignatures', () => {
-				var signKeysInDb = _.map(
-					accountRow.mem_accounts2u_multisignatures,
-					row => {
-						return row.dependentId;
-					}
-				);
-				return expect(signKeysInDb).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should set u_multimin field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.u_multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should set u_multilifetime field set on mem_accounts', () => {
-				return expect(accountRow.mem_accounts.u_multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-		});
-
-		describe('check sender account', () => {
-			var account;
-
-			before('get multisignature account', done => {
-				library.logic.account.get(
-					{ address: multisigAccount.address },
-					(err, res) => {
-						expect(err).to.be.null;
-						account = res;
-						done();
-					}
-				);
-			});
-
-			it('should have u_multisignatures field set on account', () => {
-				return expect(account.u_multisignatures).to.include(
-					signer1.publicKey,
-					signer2.publicKey
-				);
-			});
-
-			it('should have multimin field set on account', () => {
-				return expect(account.u_multimin).to.eql(
-					multisigTransaction.asset.multisignature.min
-				);
-			});
-
-			it('should have multilifetime field set on account', () => {
-				return expect(account.u_multilifetime).to.eql(
-					multisigTransaction.asset.multisignature.lifetime
-				);
-			});
-		});
-
-		describe('with another multisig transaction', () => {
-			var multisigTransaction2;
-			var signer3 = randomUtil.account();
-			var signer4 = randomUtil.account();
-
-			before('process multisignature transaction', done => {
-				var keysgroup = [`+${signer3.publicKey}`, `+${signer4.publicKey}`];
-				multisigTransaction2 = lisk.multisignature.createMultisignature(
-					multisigAccount.password,
-					null,
-					keysgroup,
-					4,
-					2
-				);
-				var sign3 = lisk.multisignature.signTransaction(
-					multisigTransaction2,
-					signer3.password
-				);
-				var sign4 = lisk.multisignature.signTransaction(
-					multisigTransaction2,
-					signer4.password
-				);
-				multisigTransaction2.signatures = [sign3, sign4];
-				library.logic.transaction.process(
-					multisigTransaction2,
-					multisigSender,
-					done
-				);
-			});
-
-			describe('from the same account', () => {
-				before('get multisignature account', done => {
-					library.logic.account.get(
-						{ address: multisigAccount.address },
-						(err, res) => {
-							multisigSender = res;
-							done();
-						}
-					);
-				});
-
-				it('should verify transaction', done => {
-					library.logic.transaction.verify(
-						multisigTransaction2,
-						multisigSender,
-						done
-					);
-				});
 			});
 		});
 	});
